requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-decouple==3.8
Pillow==10.2.0
django-allauth==0.61.1
django-rest-framework==0.1.0
channels==4.0.0
daphne==4.1.0
django-cors-headers==4.3.1
django-filter==24.1
django-storages==1.14.2
boto3==1.34.69
python-dotenv==1.0.1 